requirements.txt
ollama>=0.1.6
pydantic>=1.10.7
flask>=2.2.0
flask-cors>=3.0.10
openai>=0.27.0
rich>=13.3.0
pytest>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0